// src/types/Attachment.tsx
{/*  ===================================THIS FILE WAS AUTO GENERATED=================================== */}

export type Attachment = {
    attachmentID: string;
    entityType: string;
    entityID: string;
    fileName: string;
    mimeType: string;
    fileSizeBytes: number;
    storageProvider: string;
    storageReference: string;
    attachmentType: string;
    description: string;
    capturedDate: string;
    capturedBy: string;
};